+++ src/net/packet/handler/list_chars/reader.rs
@@ -25,7 +25,7 @@ use std::io::Cursor;
 
 #[derive(Clone)]
 pub struct ListCharsReader {
-    pub channel_id: i16,
+    pub channel_id: u8,
     pub world_id: i16,
 }
 
@@ -36,7 +36,7 @@ impl ListCharsReader {
         let skip = 1;
         pkt_reader.read_bytes(skip).map_err(ReadError)?;
         let world_id = pkt_reader.read_byte().map_err(ReadError)? as i16;
-        let channel_id = pkt_reader.read_byte().map_err(ReadError)? as i16;
+        let channel_id = pkt_reader.read_byte().map_err(ReadError)? as u8;
         Ok(Self {
             channel_id,
             world_id,
